+++ packages/kiapi-relay/src/kiapi_relay/core/__init__.py
@@ -1,5 +1,6 @@
 from ._exceptions.relay_request_error import RelayRequestError
 from ._helpers.build_relay_response import build_relay_response
+from ._helpers.get_or_create_node_id import get_or_create_node_id
 from ._helpers.new_relay_session_id import new_relay_session_id
 from ._instances.relay_registry import relay_registry
 from ._schemas.relay_file_body import RelayFileBody
@@ -8,6 +9,7 @@ from ._schemas.relay_multipart_body import RelayMultipartBody
 from ._schemas.relay_multipart_file import RelayMultipartFile
 from ._services.base_relay import BaseRelay
 from ._services.relay_runner import RelayRunner
+from ._services.single_instance import AlreadyRunningError, SingleInstanceLock
 from ._settings import RelaySettings, settings_manager
 from ._types.relay import Relay
 from ._types.relay_delivery import RelayDelivery
@@ -20,6 +22,7 @@ from ._views.relay_request import RelayRequest
 from ._views.relay_response import RelayResponse
 
 __all__ = [
+    "AlreadyRunningError",
     "BaseRelay",
     "Relay",
     "RelayDelivery",
@@ -37,7 +40,9 @@ __all__ = [
     "RelayRunner",
     "RelaySettings",
     "RelaySpecifier",
+    "SingleInstanceLock",
     "build_relay_response",
+    "get_or_create_node_id",
     "new_relay_session_id",
     "relay_registry",
     "settings_manager",