+++ packages/kiapi-relay/src/kiapi_relay/core/_services/base_relay.py
@@ -5,6 +5,7 @@ from .._types.relay_name import RelayName
 class BaseRelay(Relay):
     def __init__(self) -> None:
         self._name: RelayName | None = None
+        self._node_id: str | None = None
 
     @property
     def name(self) -> RelayName:
@@ -15,3 +16,13 @@ class BaseRelay(Relay):
     @name.setter
     def name(self, value: RelayName) -> None:
         self._name = value
+
+    @property
+    def node_id(self) -> str:
+        if self._node_id is None:
+            raise ValueError("node_id is not set")
+        return self._node_id
+
+    @node_id.setter
+    def node_id(self, value: str) -> None:
+        self._node_id = value